Update packages.html
Heya massa,

Im sorry bout the delay. I tried to implement some php functionality to display the buttons but with the button place holder already being an element that was part of the initial document i could not create a functional generation of button so i took it out and will purely implement it through PHP.Dont worry about the touching up to much please joel as you are already doing so much and being so helpful in the group that i will be doing this very shortly i just wanted to provide you with a production copy for your own library to play with before our work day is over(5 p.m XDDDDDD).